Convert CheckoutTerms to TS
